Working on mem leak issues in parser.

+++ macros.h
@@ -1,5 +1,5 @@
 
-#define BNS_ARRAY_COUNT(x) (sizeof(x)/sizeof((x)[0]))
+#define BNS_ARRAY_COUNT(x) ((int)(sizeof(x)/sizeof((x)[0])))
 
 #define BNS_UNUSED(x) (void(x))
